+++ CHANGELOG.md
@@ -18,6 +18,7 @@ SPDX-License-Identifier: MIT
     where `${XDG_CONFIG_HOME}` defaults to `${HOME}/.config`.  The
     configuration file can be used to set defaults for the command-line
     options.
+  - Add the `--config-file [file]` option to set a custom configuration file.
 
 ## v0.1.3 (2020-07-28)